requirements.txt
streamlit>=1.31.0
langchain>=0.2.6
langchain-core>=0.2.10
langchain-google-genai>=1.0.7
python-dotenv